OneDrive - MARVAL S.A/00 Extractor de datos Facturas/DocuMarval/requirements.txt
streamlit>=1.39.0,<2.0.0
google-generativeai>=0.8.0,<1.0.0
pdf2image>=1.17.0,<2.0.0
Pillow>=10.4.0,<11.0.0
pandas>=2.2.2,<3.0.0
openpyxl>=3.1.5,<4.0.0
python-dotenv>=1.0.1,<2.0.0
numpy>=1.24.0,<2.0.0
PyPDF2>=3.0.0,<4.0.0